@@ -19,20 +19,23 @@
   "license": "ISC",
   "homepage": "https://github.com/enigmatic/AdventOfCode",
   "dependencies": {
-    "@types/chai": "^4.2.6",
-    "@types/node": "^10.12.12",
-    "@types/rbush": "^2.0.2",
-    "aoc-copilot": "^1.5.9",
-    "bigint-crypto-utils": "^2.4.2",
-    "chai": "^4.2.0",
-    "chalk": "^2.4.2",
-    "prettier": "^1.19.1",
-    "rbush": "^2.0.2",
-    "rbush-knn": "^2.1.0",
+    "@types/chai": "^5.0.1",
+    "@types/node": "^22.10.1",
+    "@types/rbush": "^4.0.0",
+    "aoc-copilot": "^1.6.0",
+    "bigint-crypto-utils": "^3.3.0",
+    "chai": "^5.1.2",
+    "chalk": "^5.3.0",
+    "prettier": "^3.4.1",
+    "rbush": "^4.0.1",
+    "rbush-knn": "^4.0.0",
     "simple-pathfinder": "^1.0.2",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.11.0",
+    "ts-node": "^10.9.2",
+    "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "^3.2.1"
+    "typescript": "^5.7.2"
+  },
+  "devDependencies": {
+    "earl": "^1.3.0"
   }
 }
