@@ -24,9 +24,28 @@ async function solve(
 
   let answer: number | string = 0;
 
-  console.log(inputs);
-  console.log(additionalInfo);
-  throw new Error("Not implemented"); // <-- Replace with your solution
+  //console.log(inputs);
+  //console.log(additionalInfo);
+
+  let histA = new Array<number>();
+  let histB = new Array<number>();
+  let mapB = new Map<number, number>();
+  inputs.forEach(v => {
+    let values = v.split(/\s+/).map(v => parseInt(v));
+    histA.push(values[0]);
+    histB.push(values[1]);
+    mapB.set(values[1], (mapB.get(values[1]) || 0) + 1);
+  });
+
+  if (part === 1) {
+    histA.sort();
+    histB.sort();
+    answer = histA.reduce((a, v, i)=> a + Math.abs(v - histB[i]),0);
+  } else {
+    answer = histA.reduce((a, v)=> a + v * (mapB.get(v) || 0) ,0);
+  }
+  
+  //throw new Error("Not implemented"); // <-- Replace with your solution
   return answer;
 }
 
